+++ submit.sh
@@ -6,7 +6,6 @@ SPARK_MASTER="local[*]"
 CASSANDRA_HOST="localhost"
 
 RAW_KEYSPACE="eth_raw"
-TAG_KEYSPACE="tagpacks"
 TGT_KEYSPACE="eth_transformed"
 BUCKET_SIZE=25000
 
@@ -17,10 +16,10 @@ if [ -z "$SPARK_HOME" ] ; then
 fi
 
 EXEC=$(basename "$0")
-USAGE="Usage: $EXEC [-h] [-m MEMORY_GB] [-c CASSANDRA_HOST] [-s SPARK_MASTER] [--src_keyspace RAW_KEYSPACE] [--tag_keyspace TAG_KEYSPACE] [--tgt_keyspace TGT_KEYSPACE] [--bucket_size BUCKET_SIZE]"
+USAGE="Usage: $EXEC [-h] [-m MEMORY_GB] [-c CASSANDRA_HOST] [-s SPARK_MASTER] [--raw_keyspace RAW_KEYSPACE] [--tgt_keyspace TGT_KEYSPACE] [--bucket_size BUCKET_SIZE]"
 
 # parse command line options
-args=$(getopt -o hc:m:s: --long raw_keyspace:,tag_keyspace:,tgt_keyspace:,bucket_size:,currency: -- "$@")
+args=$(getopt -o hc:m:s: --long raw_keyspace:,tgt_keyspace:,bucket_size:,currency: -- "$@")
 eval set -- "$args"
 
 while true; do
@@ -43,10 +42,6 @@ while true; do
         ;;
         --raw_keyspace)
             RAW_KEYSPACE="$2"
-            shift 2
-        ;;
-        --tag_keyspace)
-            TAG_KEYSPACE="$2"
             shift 2
         ;;
         --tgt_keyspace)
@@ -79,7 +74,6 @@ done
 echo -en "Starting on $CASSANDRA_HOST with master $SPARK_MASTER" \
          "and $MEMORY memory ...\n" \
          "- raw keyspace:    $RAW_KEYSPACE\n" \
-         "- tag keyspace:    $TAG_KEYSPACE\n" \
          "- target keyspace: $TGT_KEYSPACE\n" \
          "- bucket size:     $BUCKET_SIZE\n"
 
@@ -91,10 +85,9 @@ echo -en "Starting on $CASSANDRA_HOST with master $SPARK_MASTER" \
   --conf spark.cassandra.connection.host="$CASSANDRA_HOST" \
   --conf spark.sql.session.timeZone=UTC \
   --conf spark.sql.extensions=com.datastax.spark.connector.CassandraSparkExtensions \
-  --packages com.datastax.spark:spark-cassandra-connector_2.12:3.1.0,org.rogach:scallop_2.12:4.0.2,joda-time:joda-time:2.10.10 \
-  target/scala-2.12/graphsense-ethereum-transformation_2.12-0.5.1.jar \
+  --packages com.datastax.spark:spark-cassandra-connector_2.12:3.1.0,org.rogach:scallop_2.12:4.1.0,joda-time:joda-time:2.10.10 \
+  target/scala-2.12/graphsense-ethereum-transformation_2.12-0.5.2.jar \
   --raw-keyspace "$RAW_KEYSPACE" \
-  --tag-keyspace "$TAG_KEYSPACE" \
   --target-keyspace "$TGT_KEYSPACE" \
   --bucket-size "$BUCKET_SIZE"