+++ CHANGELOG.md
@@ -3,11 +3,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.5.2] 2022-03-10
+### Changed
+- Improved balance calculation
+
+### Removed
+- Removed tag handling (see graphsense/graphsense-tagstore)
+
 ## [0.5.1] 2021-11-29
 ### Changed
 - Upgrade to Spark 3
 - Improved Cassandra schema
 - Changed package name
+
 ### Added
 - Added command-line arguments to spark submit script
 - Added `trace` table